requirements.txt
flask==2.3.3
redis==5.0.1
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
lightfm==1.17
faiss-cpu==1.7.4
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
scipy==1.11.1
